frontends/workflowui/backend/requirements.txt
Flask==3.1.3
Flask-CORS==6.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.32.4
pydantic==2.5.2
SQLAlchemy==2.0.23
python-dateutil==2.8.2
PyJWT==2.10.1
gunicorn==23.0.0
email-validator==2.1.0
imapclient==3.0.1
